+++ Modules/Ruleset.Initialize/Public/Initialize-Service.ps1
@@ -79,6 +79,9 @@ It should be used in conjunction with the rest of a module "Ruleset.Initialize"
 
 TODO: Some services are logged as change from ex. from Manual to Manual, but that's not change,
 this will happen ie. if restarting service.
+TODO: Service 'Function Discovery Provider Host (fdPHost)' cannot be stopped due to the following error:
+Cannot stop 'FDResPub' service on computer '.'.
+This error indicates this script fails to recognize service status that is not in valid state to be configured
 
 .LINK
 https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.Initialize/Help/en-US/Initialize-Service.md